refactor: change the list episodes view controler
